+++ src/client/modules/fetchPageContent.ts
@@ -1,20 +1,5 @@
 import { marked } from "marked";
-
-export interface PageSeed {
-  path: string;
-  indexed: boolean;
-  title?: string;
-}
-
-export interface PageContent {
-  indexed: boolean;
-  rawPath: string;
-  url: string;
-  status: number;
-  text: string;
-  title: string;
-  html: string;
-}
+import { PageSeed, PageContent } from "./types";
 
 export async function fetchPageContent({
   path,
